integrate exceptions into reactions
the idea is that if an exception is raised then we assume that we've
made a bad choice in one of the signals, unless an exception is raised
for all possible choices.  In that case, we just pick one of them and
shut everything down, causing react! to always raise that exception

Currently the choice of which exception is non-deterministic, but we
could try tracking the order of the pars and put some kind of an
ordering on signals to get the first one, possibly