adjust signal-value handling
When asking for a signal's value, it is not clear how to know which
signal emissions are "downstream" of the call to `signal-value`. So,
instead of trying to guess, let's see how far we can get by requiring
the programmer to tell us.

There is one strange implication of this choice, however. Consider
this schematic program:
```
(begin
  (par (begin (signal-value S1 #:can (set S2 S3))
              '....stuff-right-after-the-signal-value-call....)
       (present? S4))
  '...stuff-much-later-after-the-par...)
```

After this commit esterel-in-racket treats this program as if it is
this program, for the purpose of Can:

```
(begin
  (par (begin
         (emit S2)
         (emit S3)
         (pause))
       (present? S4))
  '...stuff-much-later-after-the-par...)
```

That is, there is a `(pause)` inserted in the program and Can will not
explictly see the code in "stuff-much-later-after-the-par". That might
be a problem! But, perhaps, we should try out some programs and see
how bad this is.