Fork each time a signals presence is demanded,
instead of deciding at the start of can mode how many
times to fork based on the number of signals.

The difference is illustrated by this program:

```
   (with-signal (s)
     (if (present? s)
         (void)
         (if (present? s)
             (emit s)
             (void))))

```

If you decide up front to explore twice, one with `s` set to true and
once with it set to false, then you will not see the emission, as the
two `if`s have the same test, so going into the else branch on the
first one means you cannot go into the else branch of the second one.

If, instead, you explore from the first presence test with both true
and false and then again from the second test with both true and
false, then you'll see the emission and be able to detect that this
program is nonconstructive.